# 몬스터 추가 및 스테이지 추가

+++ Assets/Scripts/Enemy/Boss/FlyingDemon/Meteor.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Meteor : MonoBehaviour
 {
-    public float atk;
+    public int atk;
     [SerializeField]
     private GameObject hitBox;
