+++ Assets/Scripts/Enemy/Stage0/Zombie.cs
@@ -58,12 +58,12 @@ public class Zombie : Enemy
 
         if (collision.gameObject.CompareTag("PlayerHitBox"))
         {
-            float damage = collision.gameObject.GetComponentInParent<Player>().playerATK;
+            int damage = collision.gameObject.GetComponentInParent<Player>().playerATK;
             DamageProcess(damage);
         }
     }
 
-    void DamageProcess(float damage)
+    void DamageProcess(int damage)
     {
         hp -= damage;
